refactor: getter 활용해보기
